pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pydatatask"
dynamic = ["version"]
authors = [
    { name = "Audrey Dutcher", email = "audrey@rhelmot.io" },
]
description = "Library for constructing data-centric processing pipelines with few invariants"
requires-python = ">=3.8"
keywords = ["pipeline", "data", "task", "repository", "processing", "async", "kubernetes", "s3", "mongodb", "docker", "ssh"]

dependencies = [
    "kubernetes",
    "kubernetes_asyncio",
    "docker-registry-client-async",
    "python-dxf",
    "jinja2",
    "pyyaml",
    "networkx",
    "aiobotocore",
    "types-aiobotocore[essential]",
    "motor",
    "ipython",
    "aiofiles",
    "aioshutil",
    "psutil",
    "asyncssh",
    "aiodocker",
    "sonyflake-py",
]

[project.optional-dependencies]
all = [
    "pydatafs"
]
fuse = [
    "pydatafs"
]
dev = [
    "tox>=4.0.0",
    "black",
    "isort",
    "pre-commit",
]
type = [
    "mypy",
    "kubernetes-typed",
    "types-pyyaml",
    "networkx-stubs",
    "types-aiofiles",
    "types-psutil",
    "pyfuse3-stubs",
]
test = [
    "pytest",
    "pytest-cov",
    "coverage[toml]"
]
lint = [
    "pylint",
]
doc = [
    "sphinx",
    "furo",
]

[project.urls]
repository = "https://github.com/rhelmot/pydatatask"
documentation = "https://pydatatask.readthedocs.io/"

[project.readme]
file = "README.md"
content-type = "text/markdown"

[project.license]
file = "LICENSE"

[tool.setuptools]
zip-safe = false
package-data = { pydatatask = ["py.typed"] }
packages = ["pydatatask"]

[tool.setuptools.dynamic]
version = {attr = "pydatatask.__version__"}

# https://github.com/tomplus/kubernetes_asyncio/issues/230
[[tool.mypy.overrides]]
module = "kubernetes_asyncio.*"
ignore_missing_imports = true

# lol
[[tool.mypy.overrides]]
module = "pydatafs.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=pydatatask --cov-report=xml -ra --log-level INFO"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
force_to_top = ["typing"]
from_first = true
skip = ["__init__.py"]

[tool.pylint]
py-version = "3.8"

[tool.pylint.FORMAT]
max-line-length = 120

[tool.pylint.'MESSAGES CONTROL']
disable = [
    "invalid-name",
    "len-as-condition",
    "locally-disabled",
    "no-else-return",
    "protected-access",
    "too-few-public-methods",
    "too-many-ancestors",
    "too-many-arguments",
    "too-many-branches",
    "too-many-instance-attributes",
    "too-many-lines",
    "too-many-locals",
    "too-many-nested-blocks",
    "too-many-public-methods",
    "too-many-return-statements",
    "too-many-statements",
    "unidiomatic-typecheck",
    "consider-using-f-string",
    "cyclic-import",
    "broad-exception-raised",  # TODO fix
    "invalid-overridden-method",  # complains about the sync/async problem with abstract methods
]

[tool.coverage.report]
exclude_lines = [
    # Have to re-enable the standard pragma
    "pragma: no cover",

    # Don't complain about missing debug-only code:
    "def __repr__",
    "if self\\.debug",

    # Don't complain if tests don't hit defensive assertion code:
    "raise AssertionError",
    "raise NotImplementedError",

    # Don't complain if non-runnable code isn't run:
    "if 0:",
    "if __name__ == .__main__.:",

    # Don't complain about abstract methods, they aren't run:
    "@(abc\\.)?abstract",
]

[tool.ruff]
line-length = 120

[tool.docformatter]
style = "sphinx"
recursive = true
pre-summary-space = true
wrap-descriptions = 120
wrap-summaries = 116
